feat(gmail): refactor polling method
